refactor: moved state machine parsing to dotlottie-rs
