+++ dotlottie-fms/src/lib.rs
@@ -5,7 +5,6 @@ mod functions;
 mod manifest;
 mod manifest_animation;
 mod manifest_themes;
-mod state_machine_parser;
 mod tests;
 mod utils;
 
@@ -16,7 +15,6 @@ pub use crate::functions::*;
 pub use crate::manifest::*;
 pub use crate::manifest_animation::*;
 pub use crate::manifest_themes::*;
-pub use crate::state_machine_parser::*;
 pub use crate::utils::*;
 
 extern crate jzon;
