+++ dotlottie-rs/src/state_machine/lib.rs
@@ -7,6 +7,7 @@ pub mod transitions;
 
 pub use crate::errors::*;
 pub use crate::events::*;
+pub use crate::parser::*;
 pub use crate::state_machine::*;
 pub use crate::states::*;
 pub use crate::transitions::*;
